Fast baseline
This establishes a good baseline (0.02 ms, so 100fps) for tweaking the pipeline. The issue we're running into is that, for some reason, the bvh branch is slow in spite of the fact that many aspects of its topology are similar. So the goal is to move this closer and closer to that until we get the sought-after slowdown.